+++ FreeAIr/Commands/InvokeLLMCommand.cs
@@ -20,7 +20,7 @@ namespace FreeAIr.Commands
         {
             await ThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync();
 
-            if (string.IsNullOrEmpty(ApiPage.Instance.Token))
+            if (!InternalPage.Instance.IsActiveAgentHasToken())
             {
                 await VS.MessageBox.ShowErrorAsync(
                     Resources.Resources.Error,
@@ -53,7 +53,7 @@ namespace FreeAIr.Commands
                     std
                     ),
                 null,
-                null
+                FreeAIr.BLogic.ChatOptions.Default
                 );
             if (chat is null)
             {