+++ FreeAIr/Commands/Other/ChooseModelCommand.cs
@@ -12,14 +12,9 @@ namespace FreeAIr.Commands.Other
 
         protected override void BeforeQueryStatus(EventArgs e)
         {
-            var uri = ApiPage.Instance.TryBuildEndpointUri();
-            if (uri is null)
-            {
-                this.Command.Enabled = false;
-                return;
-            }
-
-            if (string.Compare(uri.Host, "openrouter.ai", true) == 0)
+            var agent = InternalPage.Instance.GetActiveAgent();
+            var uri = agent.TryBuildEndpointUri();
+            if (agent.IsOpenRouterAgent())
             {
                 this.Command.Enabled = true;
             }