+++ FreeAIr/Options/OptionsProvider.cs
@@ -5,11 +5,6 @@ namespace FreeAIr
     internal partial class OptionsProvider
     {
         [ComVisible(true)]
-        public class ApiPageOptions : BaseOptionPage<ApiPage>
-        {
-        }
-
-        [ComVisible(true)]
         public class ResponsePageOptions : BaseOptionPage<ResponsePage>
         {
         }